# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Список всех постов",
                "operationId": "getPostsList",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/postData"
                                            }
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Создать новый пост",
                "operationId": "createPost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/storePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пост создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/postData"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/storePostValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить пост по ID",
                "operationId": "getPostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор поста",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/postData"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Обновить пост",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор поста",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updatePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пост обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/postData"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/updatePostValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Удалить пост",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор поста",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{id}/comments": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Комментарии поста",
                "operationId": "getPostComments",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор поста",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/commentData"
                                            }
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "500Error": {
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "type": "string",
                        "example": "Error string"
                    }
                },
                "type": "object"
            },
            "404Error": {
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "type": "string",
                        "example": "Not found"
                    }
                },
                "type": "object"
            },
...
            "storePostRequest": {
                "description": "Запрос на создание поста",
                "required": [
                    "user_id",
                    "body"
                ],
                "properties": {
                    "user_id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer",
                        "example": 1
                    },
                    "body": {
                        "description": "Текст поста",
                        "type": "string",
                        "example": "Это содержимое поста"
                    }
                },
                "type": "object"
            },
            "storePostValidationErrorResponse": {
                "description": "Ответ при ошибках валидации при создании поста",
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "properties": {
                            "user_id": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The user_id field is required."
                                }
                            },
                            "body": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The body field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "updatePostRequest": {
                "description": "Запрос на обновление поста",
                "required": [
                    "body"
                ],
                "properties": {
                    "body": {
                        "description": "Новый текст поста",
                        "type": "string",
                        "example": "Обновлённый текст поста"
                    }
                },
                "type": "object"
            },
            "updatePostValidationErrorResponse": {
                "description": "Ответ при ошибках валидации при обновлении поста",
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "properties": {
                            "body": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The body field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "commentData": {
                "description": "Данные комментария",
                "required": [
                    "id",
                    "post_id",
                    "user",
                    "body",
                    "created"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор комментария",
                        "type": "integer",
                        "example": 42
                    },
                    "post_id": {
                        "description": "Идентификатор поста",
                        "type": "integer",
                        "example": 10
                    },
                    "user": {
                        "$ref": "#/components/schemas/userData"
                    },
                    "body": {
                        "description": "Текст комментария",
                        "type": "string",
                        "example": "Отличный пост!"
                    },
                    "created": {
                        "description": "Дата и время создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-05T11:15:00Z"
                    }
                },
                "type": "object"
            },
            "postData": {
                "description": "Данные поста",
                "required": [
                    "id",
                    "user",
                    "body",
                    "created"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор поста",
                        "type": "integer",
                        "example": 10
                    },
                    "user": {
                        "$ref": "#/components/schemas/userData"
                    },
                    "body": {
                        "description": "Содержимое поста",
                        "type": "string",
                        "example": "Hello, world!"
                    },
                    "created": {
                        "description": "Дата и время создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-05T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "userData": {
                "description": "Данные пользователя",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Alice"
                    }
                },
                "type": "object"
            }
        }
    },
...
}